fix(api): drop unused SiteErrorDetails export and payload helper
knip failed on the exported type. Match the timeout-details pattern:
test getSiteErrorDetails directly and build the scrape JSON inline.
